ACTIN-2158: move resource loading into its own source file
